zenzone-backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.58.2
psycopg2-binary==2.9.9  # For PostgreSQL database
python-dotenv==1.0.0
torch==2.1.0
transformers==4.35.0
numpy==1.24.3
scikit-learn==1.3.0